Updating the case view and how the case renders
