+++ package.json
@@ -20,11 +20,12 @@
     "eslint": "8.35.0",
     "eslint-config-next": "13.2.3",
     "faker": "^5.5.3",
-    "next": "^13.2.3",
+    "next": "^13.2.4",
     "react": "^18.2.0",
     "react-big-calendar": "^1.6.9",
     "react-bootstrap": "^2.7.2",
     "react-calendar": "^4.1.1",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "sharp": "^0.32.0"
   }
 }
