+++ pages/dashboard.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const AboutPage = () => {
+  return (
+    <div>
+      <h1>About Page</h1>
+      <p>This is the about page of my Next.js application.</p>
+    </div>
+  );
+};
+
+export default AboutPage;
